Tweak random number behavior a bit
This changes the OpQuest behavior to just being more or less a coin
toss. 50/50 chances always feel rigged when you're looking for
something random, but I guess that's just random numbers for you.
I guess we don't really want random but a nice distribution of coin
tosses in either direction, which sucks, because that's not random.

This also fixes AnyChar behavior where it was generating two random
numbers and only using the first to determine if a newline was
generated. This could result in non-ASCII characters (i.e., assuming
ASCII is 7 bits) being generated.

In addition, CharClass will no longer waste time calling randint if it
selects a rune pair of the same rune. This doesn't have too much effect
on anything since randint will already return 0 if passed 0, but it
makes it a little clearer since you don't need to look at randint to
see why it works.

Change-Id: Ib08d54436dc9c82870699fa1585c4ec248673d1f