OpCharClass: don't branch on ranges
Instead of using two randint calls to branch on the rune pair and then
generate a random number inside that pair's range, generate a random
number across all ranges and then figure out which pair it's in. This
is far less efficient but it seems to produce better results than
branching on only a few pairs and that are then similarly small ranges
of runes.

Change-Id: I172789f9220512a0e570d19fbc5ac55b3cea969a