Tweak package doc a bit, fix wording
This adds a line with just the `go get` bit for people looking at the
godoc.org page and updates/trims the javascript example a bit.
Honestly, I think I just made it weirder, and now the resulting
javascript throws an error, but whatever.

The second change is also because of godoc.org (not a negative thing,
just me being picky): the example caused the <pre> element to scroll,
which looks ugly.

Adds notices about EOT/EOL handling as well as removing a word that
shouldn't be there ("run").

Adds a line about CLI options being available via regen -help.

Change-Id: Icc9f756828fcbc39563e7f0ddbed100f5c85410b